+++ app/src/main/java/com/lukasz/witkowski/training/planner/navigation/Navigation.kt
@@ -22,8 +22,8 @@ import com.lukasz.witkowski.training.planner.exercise.createExercise.CreateExerc
 import com.lukasz.witkowski.training.planner.training.createTraining.CreateTrainingScreen
 import com.lukasz.witkowski.training.planner.training.createTraining.PickExerciseScreen
 import com.lukasz.witkowski.training.planner.exercise.exercisesList.ExercisesScreen
-import com.lukasz.witkowski.training.planner.statistics.TrainingSessionScreen
-import com.lukasz.witkowski.training.planner.statistics.TrainingSessionViewModel
+import com.lukasz.witkowski.training.planner.trainingSession.TrainingSessionScreen
+import com.lukasz.witkowski.training.planner.trainingSession.TrainingSessionViewModel
 import com.lukasz.witkowski.training.planner.ui.trainingOverview.TrainingOverviewScreen
 import com.lukasz.witkowski.training.planner.ui.trainingOverview.TrainingOverviewViewModel
 import com.lukasz.witkowski.training.planner.training.trainingsList.TrainingsScreen
