+++ app/src/main/java/com/lukasz/witkowski/training/planner/trainingSession/TrainingSessionSummaryScreen.kt
@@ -1,0 +1,9 @@
+package com.lukasz.witkowski.training.planner.trainingSession
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun TrainingSessionSummaryScreen() {
+    Text(text = "Summary")
+}
