+++ app/src/main/java/com/lukasz/witkowski/training/planner/trainingSession/TrainingSessionViewModel.kt
@@ -1,4 +1,4 @@
-package com.lukasz.witkowski.training.planner.statistics
+package com.lukasz.witkowski.training.planner.trainingSession
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
@@ -11,7 +11,6 @@ import com.lukasz.witkowski.training.planner.training.domain.TrainingPlanId
 import com.lukasz.witkowski.training.planner.training.presentation.TrainingPlanMapper
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.launch
 import javax.inject.Inject
