+++ statistics/src/main/java/com/lukasz/witkowski/training/planner/statistics/presentation/DefaultTrainingSessionController.kt
@@ -1,8 +1,6 @@
 package com.lukasz.witkowski.training.planner.statistics.presentation
 
 import com.lukasz.witkowski.training.planner.statistics.application.TrainingSessionService
-import com.lukasz.witkowski.training.planner.training.presentation.TrainingExercise
-import com.lukasz.witkowski.training.planner.training.presentation.TrainingExerciseMapper
 import com.lukasz.witkowski.training.planner.training.presentation.TrainingPlan
 import com.lukasz.witkowski.training.planner.training.presentation.TrainingPlanMapper
 import kotlinx.coroutines.flow.MutableStateFlow
